Lift udp_transmit helper out of endpoint.rs
